fix(`Shop.jsx`): Fetching not using latest sortBy
The getAPIURL func was not using the latest value of the state sortBy
when it was called as the state updation of the sortBy state was
dependent on the orderBy state changing in value. However, they have to
change together for the fetching logic to work correctly. Ensure that
happens by batching the orderBy and sortBy state updates